Get retention times

Get retention times from a list of chromatograms or a peak_table object.

Usage

get_times(x, idx = 1)

Arguments

x

A list of chromatograms or peak_table object.

idx

Index of chromatogram from which to extract times.

Value

Numeric vector of retention times from the chromatogram specified by idx.

See also

Examples

data(Sa_warp)
get_times(Sa_warp)
#>   [1] 10.00 10.02 10.04 10.06 10.08 10.10 10.12 10.14 10.16 10.18 10.20 10.22
#>  [13] 10.24 10.26 10.28 10.30 10.32 10.34 10.36 10.38 10.40 10.42 10.44 10.46
#>  [25] 10.48 10.50 10.52 10.54 10.56 10.58 10.60 10.62 10.64 10.66 10.68 10.70
#>  [37] 10.72 10.74 10.76 10.78 10.80 10.82 10.84 10.86 10.88 10.90 10.92 10.94
#>  [49] 10.96 10.98 11.00 11.02 11.04 11.06 11.08 11.10 11.12 11.14 11.16 11.18
#>  [61] 11.20 11.22 11.24 11.26 11.28 11.30 11.32 11.34 11.36 11.38 11.40 11.42
#>  [73] 11.44 11.46 11.48 11.50 11.52 11.54 11.56 11.58 11.60 11.62 11.64 11.66
#>  [85] 11.68 11.70 11.72 11.74 11.76 11.78 11.80 11.82 11.84 11.86 11.88 11.90
#>  [97] 11.92 11.94 11.96 11.98 12.00 12.02 12.04 12.06 12.08 12.10 12.12 12.14
#> [109] 12.16 12.18 12.20 12.22 12.24 12.26 12.28 12.30 12.32 12.34 12.36 12.38
#> [121] 12.40 12.42 12.44 12.46 12.48 12.50 12.52 12.54 12.56 12.58 12.60 12.62
#> [133] 12.64 12.66 12.68 12.70 12.72 12.74 12.76 12.78 12.80 12.82 12.84 12.86
#> [145] 12.88 12.90 12.92 12.94 12.96 12.98 13.00 13.02 13.04 13.06 13.08 13.10
#> [157] 13.12 13.14 13.16 13.18 13.20 13.22 13.24 13.26 13.28 13.30 13.32 13.34
#> [169] 13.36 13.38 13.40 13.42 13.44 13.46 13.48 13.50 13.52 13.54 13.56 13.58
#> [181] 13.60 13.62 13.64 13.66 13.68 13.70 13.72 13.74 13.76 13.78 13.80 13.82
#> [193] 13.84 13.86 13.88 13.90 13.92 13.94 13.96 13.98 14.00 14.02 14.04 14.06
#> [205] 14.08 14.10 14.12 14.14 14.16 14.18 14.20 14.22 14.24 14.26 14.28 14.30
#> [217] 14.32 14.34 14.36 14.38 14.40 14.42 14.44 14.46 14.48 14.50 14.52 14.54
#> [229] 14.56 14.58 14.60 14.62 14.64 14.66 14.68 14.70 14.72 14.74 14.76 14.78
#> [241] 14.80 14.82 14.84 14.86 14.88 14.90 14.92 14.94 14.96 14.98 15.00 15.02
#> [253] 15.04 15.06 15.08 15.10 15.12 15.14 15.16 15.18 15.20 15.22 15.24 15.26
#> [265] 15.28 15.30 15.32 15.34 15.36 15.38 15.40 15.42 15.44 15.46 15.48 15.50
#> [277] 15.52 15.54 15.56 15.58 15.60 15.62 15.64 15.66 15.68 15.70 15.72 15.74
#> [289] 15.76 15.78 15.80 15.82 15.84 15.86 15.88 15.90 15.92 15.94 15.96 15.98
#> [301] 16.00 16.02 16.04 16.06 16.08 16.10 16.12 16.14 16.16 16.18 16.20 16.22
#> [313] 16.24 16.26 16.28 16.30 16.32 16.34 16.36 16.38 16.40 16.42 16.44 16.46
#> [325] 16.48 16.50 16.52 16.54 16.56 16.58 16.60 16.62 16.64 16.66 16.68 16.70
#> [337] 16.72 16.74 16.76 16.78 16.80 16.82 16.84 16.86 16.88 16.90 16.92 16.94
#> [349] 16.96 16.98 17.00 17.02 17.04 17.06 17.08 17.10 17.12 17.14 17.16 17.18
#> [361] 17.20 17.22 17.24 17.26 17.28 17.30 17.32 17.34 17.36 17.38 17.40 17.42
#> [373] 17.44 17.46 17.48 17.50 17.52 17.54 17.56 17.58 17.60 17.62 17.64 17.66
#> [385] 17.68 17.70 17.72 17.74 17.76 17.78 17.80 17.82 17.84 17.86 17.88 17.90
#> [397] 17.92 17.94 17.96 17.98 18.00 18.02 18.04 18.06 18.08 18.10 18.12 18.14
#> [409] 18.16 18.18 18.20 18.22 18.24 18.26 18.28 18.30 18.32 18.34 18.36 18.38
#> [421] 18.40 18.42 18.44 18.46 18.48 18.50 18.52 18.54 18.56 18.58 18.60 18.62
#> [433] 18.64 18.66