Peak Table Object

S3 objects of class peak_table store chromatographic peak data along with associated metadata and reference spectra.

Components

A peak_table object contains the following elements:

tab: The peak table itself – a data.frame of intensities in a sample x peak configuration.

pk_meta: A data.frame containing peak meta-data (e.g., the spectral component, peak number, and average retention time).

sample_meta: A data.frame of sample meta-data. Must be added using attach_metadata.

instrument_meta: A data.frame of instrumental metadata transferred automatically from the chromatograms.

ref_spectra: A data.frame of reference spectra (in a wavelength x peak configuration). Must be added using attach_ref_spectra.

args: A vector of arguments given to get_peaktable to generate the peak table.

Methods

The following methods are available:

print.peak_table()

dim.peak_table()

head.peak_table()

tail.peak_table()

row.names.peak_table()

subset.peak_table()

summary.peak_table()

[.peak_table

See also

Examples

# See get_peaktable() for creation examples
# To access peak_table components:
data(pk_tab)
pk_tab$tab          # peak table
#>           V1       V2        V3        V4       V5        V6       V7        V8
#> 119 0.000000 5.111190 0.0000000 0.5864823 1.858693 15.079233 0.000000 0.4135257
#> 121 0.000000 4.033081 0.9163944 0.5221266 0.000000  7.705485 0.000000 0.0000000
#> 122 3.368245 8.228498 0.9735932 0.0000000 1.963956 25.888893 0.000000 1.8394472
#> 458 0.000000 5.655876 1.8309616 0.0000000 1.735467 14.650605 1.459654 0.3488550
#>            V9      V10      V11      V12       V13      V14      V15       V16
#> 119  79.71546 2.637016 41.60604 1.724824 1.3743838 2.273449 36.58533 15.189655
#> 121  57.27332 1.042586 18.33105 0.000000 0.8614246 7.504094 20.56684  8.879872
#> 122  59.62042 1.417466 35.83855 0.000000 1.2422142 4.241067 31.60616 12.910084
#> 458 108.93550 4.157266 33.22952 3.405194 3.1656321 5.231752 42.99198 23.968663
#>          V17      V18      V19       V20       V21      V22       V23
#> 119 33.71629 4.186043 3.890605 1.7999191 12.435065 15.09572 2.5155590
#> 121 22.59679 1.169449 1.750044 0.1143296  6.619292 12.54686 0.7921663
#> 122 28.40685 2.232165 3.037413 0.4855545  7.976247 14.88744 2.7807659
#> 458 25.33267 5.011990 5.348641 0.2741214 12.648599 15.42859 2.3459283
#>            V24        V25      V26       V27       V28       V29       V30
#> 119 0.00000000 0.11261310 17.80024 2.0024282 0.0000000 0.7981487 0.8774407
#> 121 0.01374328 0.43874747 14.41614 0.6184847 0.5505384 0.4542899 0.5315334
#> 122 0.01900430 0.11078837 16.60867 0.9699587 1.2199894 1.2251103 0.8634827
#> 458 0.00512731 0.02601551 24.77821 1.0894581 1.0049019 1.5687582 1.7486409
#>          V31      V32      V33       V34       V35       V36       V37
#> 119 1.819662 1.023753 1.541590 0.0000000 0.2105941 0.5041651 0.0000000
#> 121 1.085835 0.000000 1.811022 0.0000000 0.5440478 0.2934669 0.0000000
#> 122 1.288960 0.000000 4.117218 0.1280478 0.5797748 0.3649663 0.3092246
#> 458 1.459723 1.647317 3.730485 0.2736762 0.7127212 0.6514810 1.2365541
#>           V38       V39        V40      V41       V42       V43        V44
#> 119  6.242450 0.1046104 0.09932078 33.73408 0.2558299 14.841595 0.00000000
#> 121  3.050482 0.0000000 0.18110972 24.36583 0.0000000  6.941289 0.03897730
#> 122 10.765649 0.3733894 0.00000000 25.07887 0.3773357 14.120471 0.03016515
#> 458  6.334782 0.0000000 0.07382435 46.69780 0.7167379 11.020983 0.53962131
#>           V45       V46       V47      V48      V49       V50       V51
#> 119 0.3332010 0.8899389 17.233670 5.787408 16.64689 0.0000000 2.2440655
#> 121 0.2602469 1.8268337  9.926202 3.454176 11.45662 0.1379431 0.6169497
#> 122 0.2959381 1.7424893 15.098178 4.611476 14.23811 0.7083038 1.2207243
#> 458 0.5018852 0.7439920 20.859029 9.333014 12.98416 0.6251158 2.9728078
#>           V52       V53      V54      V55       V56         V57         V58
#> 119 1.1409371 0.7426509 5.373707 6.566692 0.6187750 0.000000000 0.059937260
#> 121 0.4724857 0.1286407 2.819725 5.966240 0.2620172 0.002892882 0.023415605
#> 122 1.1684493 0.2942327 3.438104 7.081215 0.9350599 0.004187678 0.041697073
#> 458 1.6551325 0.3717944 5.200684 7.292224 0.8231769 0.023957666 0.005351955
#>           V59       V60       V61       V62       V63       V64       V65
#> 119  7.973773 0.1983225 0.1705325 0.3076778 0.6844486 0.5443093 0.6178561
#> 121  6.195397 0.1066713 0.2107139 0.1992947 0.9549370 0.0000000 0.4778106
#> 122  7.488124 0.2845064 0.1666236 0.0000000 0.9325107 0.0000000 1.2312923
#> 458 10.299617 0.2704478 0.3546083 0.3774084 0.5895916 0.6156869 1.3281371
pk_tab$pk_meta      # peak metadata  
#>               V1     V2     V3     V4     V5     V6     V7     V8     V9    V10
#> lambda    210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00
#> peak        1.00   2.00   3.00   4.00   5.00   6.00   7.00   8.00   9.00  10.00
#> rt         10.02  10.63  10.92  11.13  11.27  11.49  11.62  11.80  12.18  12.47
#> start      10.00  10.40  10.81  11.02  11.11  11.32  11.62  11.74  11.91  12.40
#> end        10.36  10.82  11.04  11.21  11.31  11.63  11.74  11.87  12.38  12.52
#> sd          0.07   0.07   0.10   0.04   0.05   0.05   0.07   0.06   0.04   0.05
#> tau         0.08   0.00  -0.07   0.00   0.01  -0.04  -0.03  -0.13  -0.01  -0.16
#> FWHM        0.16   0.15   0.23   0.10   0.12   0.12   0.17   0.13   0.10   0.12
#> r.squared   0.99   1.00   0.98   0.99   1.00   0.97   0.99   0.85   1.00   0.83
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V11    V12    V13    V14    V15    V16    V17    V18    V19    V20
#> lambda    210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00
#> peak       11.00  12.00  13.00  14.00  15.00  16.00  17.00  18.00  19.00  20.00
#> rt         12.69  12.93  13.16  13.47  13.76  13.95  14.44  14.75  14.99  15.26
#> start      12.54  12.86  13.03  13.30  13.53  13.84  14.23  14.62  14.90  15.16
#> end        12.91  13.04  13.27  13.53  13.84  14.09  14.58  14.84  15.12  15.30
#> sd          0.04   0.11   0.07   0.04   0.05   0.05   0.05   0.05   0.05   0.04
#> tau        -0.01  -0.21  -0.02  -0.02  -0.03  -0.03  -0.02  -0.02  -0.05  -0.12
#> FWHM        0.10   0.27   0.15   0.10   0.12   0.12   0.12   0.11   0.11   0.10
#> r.squared   1.00   0.97   0.92   0.99   0.97   0.96   0.89   0.96   0.90   0.91
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V21    V22    V23    V24    V25    V26    V27    V28    V29    V30
#> lambda    210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00 210.00
#> peak       21.00  22.00  23.00  24.00  25.00  26.00  27.00  28.00  29.00  30.00
#> rt         15.51  15.72  16.00  16.19  16.47  16.78  17.04  17.36  17.62  17.86
#> start      15.30  15.61  15.88  16.14  16.37  16.55  16.97  17.24  17.50  17.72
#> end        15.60  15.84  16.16  16.26  16.51  16.95  17.19  17.42  17.68  17.88
#> sd          0.05   0.04   0.04   0.02   0.03   0.04   0.09   0.07   0.11   0.05
#> tau        -0.01  -0.01  -0.02   0.00   0.00   0.00  -0.06  -0.12  -0.24  -0.80
#> FWHM        0.12   0.10   0.10   0.05   0.07   0.10   0.21   0.17   0.26   0.12
#> r.squared   0.98   0.78   0.97   0.70   0.99   1.00   0.83   0.88   0.92   0.91
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V31    V32    V33    V34    V35    V36    V37    V38    V39    V40
#> lambda    210.00 210.00 210.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00
#> peak       31.00  32.00  33.00   1.00   2.00   3.00   4.00   5.00   6.00   7.00
#> rt         18.06  18.24  18.50  10.22  10.55  10.88  11.26  11.49  11.72  11.90
#> start      17.92  18.12  18.20  10.18  10.39  10.70  11.06  11.24  11.65  11.79
#> end        18.11  18.28  18.53  10.38  10.70  11.04  11.30  11.66  11.83  11.94
#> sd          0.17   0.18   0.09   0.13   0.06   0.09   0.05   0.04   0.07   0.04
#> tau        -0.92  -1.23  -0.80  -0.14   0.03  -0.07   0.01  -0.04  -0.04  -0.01
#> FWHM        0.39   0.43   0.22   0.30   0.15   0.21   0.13   0.10   0.15   0.09
#> r.squared   0.86   0.83   0.94   0.98   0.98   0.98   1.00   0.98   0.95   0.93
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V41    V42    V43    V44    V45    V46    V47    V48    V49    V50
#> lambda    260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00
#> peak        8.00   9.00  10.00  11.00  12.00  13.00  14.00  15.00  16.00  17.00
#> rt         12.18  12.44  12.72  13.01  13.16  13.45  13.76  13.96  14.35  14.58
#> start      11.93  12.41  12.59  12.89  13.06  13.30  13.53  13.99  14.15  14.53
#> end        12.43  12.53  12.92  13.02  13.24  13.51  13.86  14.20  14.44  14.65
#> sd          0.04   0.10   0.08   0.04   0.14   0.05   0.05   0.05   0.04   0.13
#> tau         0.00  -0.16  -0.17   0.00 -15.28  -0.05  -0.01   0.05  -0.02  -0.30
#> FWHM        0.10   0.23   0.20   0.09   0.33   0.11   0.11   0.11   0.09   0.30
#> r.squared   1.00   0.94   0.99   0.97   0.75   0.97   0.99   0.94   0.99   0.55
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V51    V52    V53    V54    V55    V56    V57    V58    V59    V60
#> lambda    260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00 260.00
#> peak       18.00  19.00  20.00  21.00  22.00  23.00  24.00  25.00  26.00  27.00
#> rt         14.75  14.95  15.26  15.54  15.75  16.00  16.23  16.43  16.78  17.10
#> start      14.64  14.87  15.15  15.40  15.62  15.89  16.15  16.34  16.53  17.04
#> end        14.84  15.12  15.35  15.61  15.89  16.15  16.30  16.51  17.02  17.19
#> sd          0.04   0.06   0.06   0.05   0.04   0.04   0.02   0.03   0.04   0.09
#> tau        -0.02   0.02  -0.01  -0.01  -0.01  -0.01  -0.01   0.00  -0.01  -0.23
#> FWHM        0.10   0.15   0.14   0.11   0.09   0.09   0.04   0.06   0.10   0.22
#> r.squared   0.96   0.93   0.93   0.84   0.98   0.97   0.97   1.00   1.00   0.75
#> purity        NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>              V61    V62    V63    V64    V65
#> lambda    260.00 260.00 260.00 260.00 260.00
#> peak       28.00  29.00  30.00  31.00  32.00
#> rt         17.35  17.66  18.00  18.21  18.48
#> start      17.26  17.50  17.69  18.08  18.20
#> end        17.40  17.70  18.06  18.29  18.54
#> sd          0.15   0.13   0.11   0.28   0.17
#> tau        -1.31  -0.54  -0.48  -0.85  -0.28
#> FWHM        0.35   0.31   0.26   0.65   0.40
#> r.squared   0.75   0.94   0.94   0.70   0.84
#> purity        NA     NA     NA     NA     NA
pk_tab$sample_meta  # sample metadata
#> [1] NA