Function to cluster peaks by spectral similarity.

Function to cluster peaks by spectral similarity. A representative spectrum is selected for each peak in the provided peak table and used to construct a distance matrix based on spectral similarity (pearson correlation) between peaks. Currently, representative spectrum is just selected from the chromatogram with the highest absorbance at lambda max. Hierarchical clustering with bootstrap resampling is performed on the resulting correlation matrix, as implemented in the pvclust package. Bootstrap values can be used to select clusters that exceed a certain confidence threshold as defined by alpha.

cluster_spectra(
  peak_table,
  chrom_list = NULL,
  peak_no = c(5, 100),
  alpha = 0.95,
  nboot = 1000,
  plot_dend = T,
  plot_spectra = TRUE,
  verbose = TRUE,
  save = TRUE,
  parallel = TRUE,
  max.only = FALSE,
  ...
)

Arguments

peak_table

Peak table from get_peaktable.

chrom_list

A list of chromatograms in matrix form (timepoints x wavelengths).

peak_no

Minimum and maximum thresholds for the number of peaks a cluster may have.

alpha

Confidence threshold for inclusion of cluster.

nboot

Number of bootstrap replicates for pvclust.

plot_dend

Logical. If TRUE, plots dendrogram with bootstrap values.

plot_spectra

Logical. If TRUE, plots overlapping spectra for each cluster.

verbose

Logical. If TRUE, prints progress report to console.

save

Logical. If TRUE, saves pvclust object to current directory.

parallel

Logical. If TRUE, use parallel processing for pvclust.

max.only

Logical. If TRUE, returns only highest level for nested dendrograms.

...

Additional arguments to pvclust.

Value

Returns S4 "cluster" object with the following components:

peaks

a character vector containing the names of all peaks contained in the given cluster.

pval

a numeric vector of length 1 containing the bootstrap p-value (au) for the given cluster.

References

R. Suzuki, H. Shimodaira: Pvclust: an R package for assessing the uncertainty in hierarchical clustering. Bioinformatics, 22-12:1540-1542 (2006). doi: 10.1093/bioinformatics/btl117 .

Author

Ethan Bass

Examples

if (FALSE) {
cl <- cluster_spectra(pk_tab, nboot=100, max.only = F, save = F, alpha = .97)
}