Combine peaks in peak table

Utility function to combine duplicate peaks in peak table, i.e. peaks that were integrated at more than one wavelength or component. Specify number of digits to match retention time and minimum spectral correlation for a match.

combine_peaks(peak_table, tol = 0.01, min.cor = 0.9, choose = "max")

Arguments

peak_table

Peak table from get_peaktable.

tol

Tolerance for matching retention times.

min.cor

Minimum spectral correlation to confirm a match.

choose

If "max" will retain peak with highest intensity. Otherwise, the first column in the data.frame will be retained.

Value

A peak table similar to the input peak table, but with duplicate columns combined according to the specified criteria.

See also

Author

Ethan Bass

Examples

if (FALSE) {
combine_peaks(pk_tab, tol = .02, min.cor = .9)}