Correct peak positions according to a ptw warping model

Once an appropriate warping model has been established, corrected retention times can be predicted for each peak. These are stored in a separate column in the list of peak tables.

correct_peaks(peak_list, mod_list)

Arguments

peak_list

A nested list of peak tables: the first level is the sample, and the second level is the component. Every component is described by a matrix where every row is one peak, and the columns contain information on retention time, full width at half maximum (FWHM), peak width, height, and area.

mod_list

A list of ptw models.

Value

The input list of peak tables is returned with extra columns containing the corrected retention time.

See also

Author

Ron Wehrens