Correct retention time

Corrects retention time differences using parametric time warping as implemented in ptw.

correct_rt(
  chrom_list,
  models = NULL,
  lambdas = NULL,
  reference = "best",
  what = c("models", "corrected.values"),
  init.coef = c(0, 1, 0),
  n.traces = NULL,
  n.zeros = 0,
  scale = TRUE,
  trwdth = 200,
  ...
)

Arguments

chrom_list

List of matrices containing concentration profiles.

models

List of models to warp by.

lambdas

Select wavelengths to use by name.

reference

Index of the sample that is to be considered the reference sample.

what

What to return: either the 'corrected.values' (useful for visual inspection) or the warping 'models' (for further programmatic use).

init.coef

Starting values for the optimization.

n.traces

Number of traces to use.

n.zeros

Number of zeros to add.

scale

Logical. If true, scale chromatograms before warping.

trwdth

width of the triangle in the WCC criterion.

...

Optional arguments for the ptw function. The only argument that cannot be changed is warp.type: this is always equal to "global".

Value

A list of `ptw` objects or a list of warped absorbance profiles.

Note

Adapted from correctRT function in the alsace package by Ron Wehrens.

References

Eilers, P.H.C. 2004. Parametric Time Warping. Anal. Chem. 76:404-411. doi: 10.1021/ac034800e .

Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015. Fast parametric time warping of peak lists. Bioinformatics 31:3063-3065. doi: 10.1093/bioinformatics/btv299 .

Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in LC–DAD using multivariate curve resolution: the alsace package for R. Metabolomics 11:1:143-154. doi: 10.1007/s11306-014-0683-5

See also

Author

Ethan Bass

Examples

if (FALSE) { # interactive()
data(Sa_pr)
warping.models <- correct_rt(Sa_pr, what = "models", lambdas=c("210"))
warp <- correct_rt(chrom_list = Sa_pr, models = warping.models)
}