Filter peak lists

Utility function to remove peaks from a peak list, e.g. because their intensity is too low. Currently one can filter on peak height, peak area, and width at half maximum.

filter_peaks(peak_list, min_height, min_area, min_sd, max_sd)

Arguments

peak_list

A peak_list object, consisting of a nester list of peak tables, where the first level is the sample, and the second level is the spectral component. Every component is described by a matrix where every row is one peak, and the columns contain information on retention time, full width at half maximum (FWHM), peak width, height, and area.

min_height

Minimum peak height.

min_area

Minimum peak area.

min_sd

Minimal standard deviation.

max_sd

Maximum standard deviation.

Value

A peak list similar to the input peak_list, but with all rows removed from the peak tables that are not satisfying the criteria.

See also

Author

Ron Wehrens, Ethan Bass