Find peaks in chromatographic profile

Find peaks in chromatographic profile.

find_peaks(
  y,
  smooth_type = "gaussian",
  smooth_window = 1,
  smooth_width = 0.1,
  slope_thresh = 0,
  amp_thresh = 0,
  bounds = TRUE
)

Arguments

y

response (numerical vector)

smooth_type

Type of smoothing. (Defaults to "gaussian").

smooth_window

Window for smoothing. (Defaults to 1).

smooth_width

Width for smoothing. (Defaults to 0.1).

slope_thresh

Minimum threshold for peak slope. (Defaults to 0).

amp_thresh

Minimum threshold for peak amplitude. (Defaults to 0).

bounds

Logical. If TRUE, includes peak boundaries in data.frame. (Defaults to TRUE).

Value

If bounds == T, returns a data.frame containing the center, start, and end of each identified peak. Otherwise, returns a numeric vector of peak centers. All locations are expressed as indices.

Details

Find peaks with function find_peaks by looking for zero-crossings in the smoothed first derivative of a signal that exceed a given slope threshold.

Note

The find_peaks function is adapted from matlab code in Prof. Tom O'Haver's Pragmatic Introduction to Signal Processing.

References

O'Haver, Tom. Pragmatic Introduction to Signal Processing: Applications in scientific measurement. /hrefhttps://terpconnect.umd.edu/~toh/spectrum/ (Accessed January, 2022).

See also

Author

Ethan Bass

Examples

data(Sa_pr)
find_peaks(Sa_pr[[1]][,"220"])
#>    pos lower upper
#> 1   28    15    46
#> 2   49    46    51
#> 3   60    54    61
#> 4   69    61    78
#> 5   81    78    84
#> 6   89    85    93
#> 7  104    93   114
#> 8  115   114   118
#> 9  120   118   123
#> 10 131   123   139
#> 11 141   139   146
#> 12 155   150   160
#> 13 167   161   171
#> 14 182   171   188
#> 15 193   188   200
#> 16 215   200   228
#> 17 233   228   238
#> 18 243   238   254
#> 19 261   254   264
#> 20 272   264   279
#> 21 283   279   293
#> 22 297   293   303
#> 23 305   303   308
#> 24 319   316   323
#> 25 336   323   347
#> 26 351   347   363
#> 27 378   363   383
#> 28 388   383   390
#> 29 395   390   401
#> 30 408   401   411
#> 31 417   411   421