Fit chromatographic peaks to an exponential-gaussian hybrid or gaussian profile

Fit peak parameters using exponential-gaussian hybrid or gaussian function.

Usage

fit_peaks(
  y,
  pos = NULL,
  sd.max = 50,
  fit = c("egh", "gaussian"),
  max.iter = 1000,
  ...
)

Arguments

y

response (numerical vector)

pos

Locations of peaks in vector y. If NULL, find_peaks will run automatically to find peak positions.

sd.max

Maximum width (standard deviation) for peaks. Defaults to 50.

fit

Function for peak fitting. (Currently "gaussian" and exponential gaussian hybrid ("egh") are supported. Defaults to egh.)

max.iter

Maximum number of iterations to use in nonlinear least squares peak-fitting. (Defaults to 1000).

...

Additional arguments to find_peaks.

Value

Function fit_peaks returns a matrix, whose columns contain the following information:

rt

location of the maximum of the peak (x)

start

start of peak (only included in table if `bounds==TRUE`)

end

end of peak (only included in table if `bounds==TRUE`)

sd

width of the peak (x)

tau

tau parameter (only included in table if `fit=="egh"`)

FWHM

full width at half maximum (x)

height

height of the peak (y)

area

peak area

r.squared

r-squared value for linear fit of model to data.

Again, the first five elements (rt, start, end, sd and FWHM) are expressed as indices, so not in terms of the real retention times. The transformation to "real" time is done in function get_peaks.

Details

Peak parameters are calculated using fit_peaks, which fits the data to a gaussian or exponential-gaussian hybrid curve using non-linear least squares estimation as implemented in nlsLM. Area under the fitted curve is estimated using trapezoidal estimation.

Note

The fit_peaks function is adapted from Dr. Robert Morrison's DuffyTools package as well as code published in Ron Wehrens' alsace package.

References

Lan, K. & Jorgenson, J. W. A hybrid of exponential and gaussian functions as a simple model of asymmetric chromatographic peaks. Journal of Chromatography A 915, 1-13 (2001). Naish, P. J. & Hartwell, S. Exponentially Modified Gaussian functions - A good model for chromatographic peaks in isocratic HPLC? Chromatographia 26, 285-296 (1988).

See also

Author

Ethan Bass

Examples

data(Sa_pr)
fit_peaks(Sa_pr[[1]][,"220"])
#>           rt start end        sd           tau      FWHM      height
#> 1   30.69582    15  46 6.2876089    1.20032657 14.775881   8.9910726
#> 2   50.87252    46  51 0.8609421  -11.40884987  2.023214   0.3652505
#> 3   61.44532    54  61 3.2042224    2.37308316  7.529923  12.7874585
#> 4   69.97526    61  78 2.7067311   -1.01879444  6.360818 118.1461826
#> 5   82.39452    78  84 1.3890785   -0.62739645  3.264334   1.5745430
#> 6   89.64284    85  93 1.5296385    0.61202489  3.594650   2.3360448
#> 7  104.49707    93 114 2.1826134   -0.19394681  5.129142 878.3558848
#> 8   81.85278   114 118 2.4020261   22.61826109  5.644761   9.6411867
#> 9  119.85327   118 123 5.2832726  -13.86056118 12.415691   8.5707183
#> 10 131.85675   123 139 1.8315236    0.11513595  4.304081 418.7796260
#> 11 141.10661   139 146 4.0598498   -4.07497282  9.540647   8.5085781
#> 12 156.08509   150 160 1.8010645    0.01466272  4.232502   5.8543152
#> 13 167.73853   161 171 1.6517982   -0.14556475  3.881726  25.5505898
#> 14 183.22815   171 188 2.0807233   -0.22145458  4.889700 350.1547326
#> 15 193.98654   188 200 2.4682084   -1.29245509  5.800290 119.6729349
#> 16 216.44045   200 228 1.8375563   -0.13007404  4.318257 319.6260534
#> 17 234.50440   228 238 1.9619252   -0.17648210  4.610524  35.8686113
#> 18 243.58113   238 254 4.9483855    0.66568411 11.628706  15.8252538
#> 19 262.00099   254 264 1.8302984   -0.98597517  4.301201  17.3957395
#> 20 272.87432   264 279 3.0837104    0.05107111  7.246719  96.9289225
#> 21 284.45995   279 293 1.8712197   -0.34947361  4.397366 121.3286776
#> 22 298.28530   293 303 2.0217365   -0.90317975  4.751081  14.6884289
#> 24 320.15708   316 323 1.2587749    0.34041774  2.958121   0.6212100
#> 25 336.89949   323 347 2.2339189   -0.82019091  5.249709 157.2481960
#> 26 351.82166   347 363 4.3854006    2.73641782 10.305691   7.1109632
#> 27 379.01651   363 383 9.6193330  -13.36671547 22.605433   3.8470521
#> 28 389.96461   383 390 0.4505309  -11.19848619  1.058748   5.0689224
#> 29 396.32753   390 401 7.3086176   -2.14828148 17.175251   6.6276427
#> 30 410.94946   401 411 1.7810651 -123.44164907  4.185503   5.5950637
#> 31 419.42472   411 421 4.8293969  -14.83514533 11.349083   7.4589610
#>           area r.squared
#> 1   141.469804 0.9861795
#> 2     1.534341 0.2835405
#> 3    44.203814 0.9996774
#> 4   844.468381 0.9954198
#> 5     7.861763 0.9445200
#> 6     9.913595 0.9974965
#> 7  4895.362421 0.9991512
#> 8    37.726827 0.8292605
#> 9    36.686028 0.9637970
#> 10 2149.825149 0.9991802
#> 11   42.795027 0.9849679
#> 12   31.783431 0.9955796
#> 13  116.325582 0.9997472
#> 14 2159.863517 0.9987345
#> 15  749.502596 0.9870224
#> 16 1996.653868 0.9988400
#> 17  231.767485 0.9889112
#> 18  148.058990 0.8325458
#> 19   88.004196 0.9912112
#> 20  713.899661 0.9118992
#> 21  753.063437 0.9843055
#> 22   84.486071 0.9455491
#> 24    2.923790 0.9886961
#> 25  972.579140 0.9950354
#> 26   70.695957 0.7701393
#> 27   63.752468 0.9955775
#> 28   30.677295 0.9392470
#> 29   64.992481 0.8411157
#> 30   54.963920 0.8662737
#> 31   67.504644 0.8950062