Find and fit peaks from a list of chromatograms.

Finds and fits peaks and extracts peak parameters from a list of chromatograms at the specified wavelengths.

get_peaks(
  chrom_list,
  lambdas,
  fit = c("egh", "gaussian"),
  sd.max = 50,
  max.iter = 100,
  ...
)

Arguments

chrom_list

A list of profile matrices, each of the same dimensions (timepoints x wavelengths).

lambdas

Character vector of wavelengths to find peaks at.

fit

What type of fit to use. Current options are gaussian and exponential gaussian hybrid (egh).

sd.max

Maximum width (standard deviation) for peaks. Defaults to 50.

max.iter

Maximum number of iterations for non-linear least squares in fit_peaks.

...

Additional arguments to find_peaks.

Value

The result is an S3 object of class `peak_list`, containing a nested list of data.frames containing information about the peaks fitted for each chromatogram at each specified wavelength. The data.frame includes information about the retention time (`rt`), `start` and `end` of each peak, as well as the standard deviation (`sd`), `tau` (if `egh` is selected), full width at half maximum (`FWHM`), `height`, `area`, and `r.squared` (coefficient of determination). This last parameter is determined from a linear model of the fitted peak values to the raw data.

Details

Peaks are located by finding zero-crossings in the smoothed first derivative of the specified chromatographic traces (function find_peaks). At the given positions, an exponential-gaussian hybrid (or regular gaussian) function is fit to the signal using fit_peaks). The area is then calculated using a trapezoidal approximation.

Note

Function is adapted from the getAllPeaks function authored by Ron Wehrens.

References

Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in LC–DAD using multivariate curve resolution: the alsace package for R. Metabolomics 11:1:143-154. doi: 10.1007/s11306-014-0683-5 .

Author

Ethan Bass

Examples

if (FALSE) { # interactive()
data(Sa_pr)
pks <- get_peaks(Sa_pr, lambdas = c('210'), sd.max=50, fit="egh")
}