Convert peak list into an ordered peak table

Function returns a peak_table object. The first slot contains a matrix of intensities, where rows correspond to samples and columns correspond to aligned features. The rest of the slots contain various meta-data about peaks, samples, and experimental settings.

Usage

get_peaktable(
  peak_list,
  chrom_list = NULL,
  response = c("area", "height"),
  use.cor = FALSE,
  hmax = 0.2,
  plot_it = FALSE,
  ask = plot_it,
  clust = c("rt", "sp.rt"),
  sigma.t = NULL,
  sigma.r = 0.5,
  deepSplit = FALSE,
  out = c("data.frame", "matrix")
)

Arguments

peak_list

A nested list of peak tables: the first level is the sample, and the second level is the component. Every component is described by a matrix where every row is one peak, and the columns contain information on retention time, full width at half maximum (FWHM), peak width, height, and area.

chrom_list

A list of chromatographic matrices.

response

An indicator whether peak area or peak height is to be used as intensity measure. Default is peak area.

use.cor

Logical, indicating whether to use corrected retention times (by default) or raw retention times (not advised!).

hmax

Height at which the complete linkage dendrogram will be cut. Can be interpreted as the maximal inter-cluster retention time difference.

plot_it

Logical. If TRUE, for every component a stripplot will be shown indicating the clustering.

ask

Logical. Ask before showing new plot?

clust

Specify whether to perform hierarchical clustering based on spectral similarity and retention time ("sp.rt") or retention time alone ("rt"). Defaults to "rt". The "sp.rt" option is experimental and should be used with some caution.

sigma.t

Width of gaussian in retention time distance function. Controls weight of retention time.

sigma.r

Width of gaussian in spectral similarity function. Controls weight of spectral correlation.

deepSplit

Logical. Controls sensitivity to cluster splitting. If TRUE, will return more smaller clusters. See documentation for cutreeDynamic.

out

Specify "data.frame" or "matrix" as output. Defaults to `data.frame`.

Value

The function returns a peak_table object, consisting of the following elements: `tab`: the peak table itself -- a data-frame of intensities in a sample x peak configuration. peaks (columns). `pk_meta`: A data.frame containing peak meta-data (e.g. the spectral component, peak number, and average retention time). `sample_meta`: A data.frame of sample meta-data. Must be added using attach_metadata) `ref_spectra`: A data.frame of reference spectra (in a wavelength x peak configuration). Must be added using attach_ref_spectra`args`: A vector of arguments given to get_peaktable.

Details

The function performs a complete linkage clustering of retention times across all samples, and cuts at a height given by the user (which can be understood as the maximal inter-cluster retention time difference) in the simple case based on retention times. Clustering can also incorporate information about spectral similarity using a distance function adapted from Broeckling et al., 2014: $$\exp({-\frac{(1-c_{ij})^2}{2\sigma_r^2}})*\exp({-\frac{(1-(t_i - t_j)^2}{2\sigma_t^2}})$$.

If two peaks from the same sample are assigned to the same cluster, a warning message is printed to the console. These warnings can usually be ignored, but one could also consider reducing the hmax variable. However, this may lead to splitting of peaks across multiple clusters. Another option is to filter the peaks by intensity to remove small features.

Note

Adapted from https://github.com/rwehrens/alsace/blob/master/R/getPeakTable.RgetPeakTable function in the alsace package by Ron Wehrens.

References

Broeckling, C. D., F. A. Afsar, S. Neumann, A. Ben-Hur, and J. E. Prenni. 2014. RAMClust: A Novel Feature Clustering Method Enables Spectral-Matching-Based Annotation for Metabolomics Data. Anal. Chem. 86:6812-6817.

See also

Author

Ethan Bass

Examples

data(Sa_pr)
pks <- get_peaks(Sa_pr, lambdas = c('210'))
get_peaktable(pks, response = "area")
#> Warning: More than one peak of one injection in the same cluster
#> for component 1.
#> Keeping only the most intense one.
#>           X1       X2       X3       X4        X5        X6         X7       X8
#> 119 260.9597  0.00000 30.92981 94.03383  769.4859  19.92870 4051.40350    0.000
#> 121 204.0379  0.00000 45.97627 23.82418    0.0000 389.30453    0.00000    0.000
#> 122 410.5215 48.67966 29.39020 98.19779 1294.4447  91.97236    0.00000 2981.014
#> 458 295.2217  0.00000 94.02964  0.00000  243.9648 741.83980   18.46819    0.000
#>             X9        X10        X11       X12       X13       X14       X15
#> 119  133.00682 2107.71843   78.53709  68.14431 116.36674 1848.3052  759.3063
#> 121 2863.55673   44.47693  912.78961   0.00000  42.84824  375.1831    0.0000
#> 122   55.55018   70.87332 1791.92740  61.76002 211.86240    0.0000 1580.3078
#> 458 5519.70162  207.40679 1673.15656 168.58992  68.12473  280.8577    0.0000
#>           X16      X17        X18       X19       X20       X21       X22
#> 119    0.0000 1695.406    0.00000 208.38958  69.35061 198.25494 96.610047
#> 121 1016.7917    0.000 1111.87828   0.00000  56.18324  86.16721  5.486941
#> 122  645.5042 1420.343   65.72994 111.60827   0.00000 151.87067 24.277725
#> 458 2121.5831    0.000 1242.67116  22.58176 244.87533 260.53421 13.762594
#>          X23      X24      X25      X26        X27       X28      X29       X30
#> 119 631.6451 739.0477   0.0000 121.2293  0.0000000 5.6158870 2.478254  884.5136
#> 121   0.0000 332.7662 600.6048   0.0000 38.8425554 0.6490484 2.676501  698.5696
#> 122 397.3674 130.5012 720.9317 139.0383  0.7108261 0.0000000 5.538164  830.4333
#> 458   0.0000 612.7060 741.7732 112.1529  0.0000000 0.0000000 1.658587 1179.0521
#>          X31      X32      X33      X34      X35      X36       X37
#> 119 98.06441 39.18598 45.47252 84.52052 51.17507 77.06908 114.02089
#> 121 28.76913 27.29513 26.30663 22.16925 54.14121  0.00000  85.32191
#> 122 48.49794 60.99947 61.25551 43.17413 64.44801  0.00000 205.86090
#> 458 54.02170 49.80864 71.31631 78.53060 72.78320  0.00000 186.14514