Plot spectrum from peak table

Plots the trace and/or spectrum for a given peak in peak table.

# S3 method for peak_table
plot(
  x,
  ...,
  loc = NULL,
  chrom_list = NULL,
  what = "peak",
  chr = "max",
  lambda = "max",
  plot_spectrum = TRUE,
  plot_trace = TRUE,
  box_plot = FALSE,
  vars = NULL,
  spectrum_labels = TRUE,
  scale_spectrum = FALSE,
  export_spectrum = FALSE,
  verbose = TRUE
)

Arguments

x

The peak table (output from get_peaktable function).

...

Additional arguments.

loc

The name of the peak or retention time for which you wish to extract spectral data.

chrom_list

A list of chromatograms in matrix form (timepoints x wavelengths).

what

What to look for. Either "peak" to extract spectral information for a certain peak, "rt" to scan by retention time, or "click" to manually select retention time by clicking on the chromatogram. Defaults to "peak".

chr

Numerical index of chromatogram you wish to plot; "max" to plot the chromatogram with the largest signal; or "all" to plot spectra for all chromatograms.

lambda

The wavelength you wish to plot the trace at if plot_chrom ==T and/or the wavelength to be used for the determination of signal abundance.

plot_spectrum

Logical. If TRUE, plots the spectrum of the chosen peak. Defaults to TRUE.

plot_trace

Logical. If TRUE, plots the trace of the chosen peak at lambda. Defaults to TRUE.

box_plot

Logical. If TRUE, plots box plot using categories defined by vars.

vars

Independent variables for boxplot.

spectrum_labels

Logical. If TRUE, plots labels on maxima in spectral plot. Defaults to TRUE.

scale_spectrum

Logical. If TRUE, scales spectrum to unit height. Defaults to FALSE.

export_spectrum

Logical. If TRUE, exports spectrum to console. Defaults to FALSE.

verbose

Logical. If TRUE, prints verbose output to console. Defaults to TRUE.

Details

Can be used to confirm the identity of a peak or check that a particular column in the peak table represents a single compound. Can also be used to create simple box-plots to examine the distribution of a peak with respect to variables defined in sample metadata.

Author

Ethan Bass