Plot spectrum from peak table

Plots the trace and/or spectrum for a given peak in peak.table object, or plots the spectrum a particular retention time for a given chromatogram.

Usage

plot_spectrum(
  loc,
  peak_table,
  chrom_list,
  chr = "max",
  lambda = "max",
  plot_spectrum = TRUE,
  plot_trace = TRUE,
  spectrum_labels = TRUE,
  scale_spectrum = FALSE,
  export_spectrum = FALSE,
  verbose = TRUE,
  what = c("peak", "rt", "click"),
  ...
)

Arguments

loc

The name of the peak or retention time for which you wish to extract spectral data.

peak_table

The peak table (output from get_peaktable function).

chrom_list

A list of chromatograms in matrix form (timepoints x wavelengths).

chr

Numerical index of chromatogram you wish to plot, or "max" to automatically plot the chromatogram with the largest signal.

lambda

The wavelength you wish to plot the trace at if plot_trace == TRUE and/or the wavelength to be used for the determination of signal abundance.

plot_spectrum

Logical. If TRUE, plots the spectrum of the chosen peak. Defaults to TRUE.

plot_trace

Logical. If TRUE, plots the trace of the chosen peak at lambda. Defaults to TRUE.

spectrum_labels

Logical. If TRUE, plots labels on maxima in spectral plot. Defaults to TRUE.

scale_spectrum

Logical. If TRUE, scales spectrum to unit height. Defaults to FALSE.

export_spectrum

Logical. If TRUE, exports spectrum to console. Defaults to FALSE.

verbose

Logical. If TRUE, prints verbose output to console. Defaults to TRUE.

what

What to look for. Either "peak" to extract spectral information for a certain peak, "rt" to scan by retention time, or "click" to manually select retention time by clicking on the chromatogram. Defaults to "peak" mode.

...

Additional arguments.

Details

Can be used to confirm the identity of a peak or check that a particular column in the peak table represents a single compound. Retention times can also be selected by clicking on the plotted trace if what == 'click'.

Author

Ethan Bass

Examples

if (FALSE) { # interactive()
data(Sa)
pks <- get_peaks(Sa,lambda="220.00000")
pk_tab <- get_peaktable(pks)
par(mfrow=c(2,1))
plot_spectrum(loc = "X10", peak_table = pk_tab, what="peak")
}